Move generateID to zfs methond controlling mock and real implementation
