+++ internal/machines/helpers.go
@@ -3,11 +3,8 @@ package machines
 import (
 	"context"
 	"fmt"
-	"math/rand"
 	"sort"
 	"strings"
-	"sync"
-	"time"
 
 	"github.com/ubuntu/zsys/internal/i18n"
 	"github.com/ubuntu/zsys/internal/log"
@@ -145,19 +142,4 @@ func sortedStateKeys(m map[string]*State) []string {
 	}
 	sort.Strings(keys)
 	return keys
-}
-
-var seedOnce = sync.Once{}
-
-// generateID with n ascii or digits, lowercase, characters
-func generateID(n int) string {
-	seedOnce.Do(func() { rand.Seed(time.Now().UnixNano()) })
-
-	var allowedRunes = []rune("abcdefghijklmnopqrstuvwxyz0123456789")
-
-	b := make([]rune, n)
-	for i := range b {
-		b[i] = allowedRunes[rand.Intn(len(allowedRunes))]
-	}
-	return string(b)
 }
